removed redundant PackagePluginExtension that was replaced with GradlePackagerSettings
