+++ src/main/java/io/github/fvarrui/javapackager/gradle/GradlePackagerSettings.java
@@ -1,14 +1,15 @@
 package io.github.fvarrui.javapackager.gradle;
 
 import groovy.lang.Closure;
-import io.github.fvarrui.javapackager.model.MacConfig;
-import io.github.fvarrui.javapackager.model.Manifest;
-import io.github.fvarrui.javapackager.model.Scripts;
-import io.github.fvarrui.javapackager.model.WindowsConfig;
+import io.github.fvarrui.javapackager.model.*;
 import io.github.fvarrui.javapackager.packagers.Context;
 import io.github.fvarrui.javapackager.packagers.Packager;
 import io.github.fvarrui.javapackager.packagers.PackagerFactory;
 import io.github.fvarrui.javapackager.packagers.PackagerSettings;
+
+import java.io.File;
+import java.util.ArrayList;
+import java.util.HashMap;
 
 import static io.github.fvarrui.javapackager.utils.ObjectUtils.defaultIfBlank;
 import static io.github.fvarrui.javapackager.utils.ObjectUtils.defaultIfNull;
@@ -18,6 +19,39 @@ public class GradlePackagerSettings extends PackagerSettings {
 
     public GradlePackagerSettings(PackageTask packageTask) {
         this.packageTask = packageTask;
+        this.platform = Platform.auto;
+        this.additionalModules = new ArrayList<>();
+        this.additionalModulePaths = new ArrayList<>();
+        this.additionalResources = new ArrayList<>();
+        this.administratorRequired = false;
+        this.assetsDir = new File(packageTask.getProject().getProjectDir(), "assets");
+        this.bundleJre = true;
+        this.copyDependencies = true;
+        this.createTarball = false;
+        this.createZipball = false;
+        this.customizedJre = true;
+        this.description = packageTask.getProject().getDescription();
+        this.extra = new HashMap<>();
+        this.generateInstaller = true;
+        this.jreDirectoryName = "jre";
+        this.linuxConfig = new LinuxConfig();
+        this.macConfig = new MacConfig();
+        this.manifest = new Manifest();
+        this.modules = new ArrayList<>();
+        this.name = packageTask.getProject().getName();
+        this.organizationEmail = "";
+        this.useResourcesAsWorkingDir = true;
+        this.vmArgs = new ArrayList<>();
+        this.winConfig = new WindowsConfig();
+        this.outputDirectory = packageTask.getProject().getBuildDir();
+        this.scripts = new Scripts();
+        this.forceInstaller = false;
+    }
+
+    public LinuxConfig linuxConfig(Closure<LinuxConfig> closure) {
+        linuxConfig = new LinuxConfig();
+        packageTask.getProject().configure(linuxConfig, closure);
+        return linuxConfig;
     }
 
     public MacConfig macConfig(Closure<MacConfig> closure) {
@@ -47,7 +81,7 @@ public class GradlePackagerSettings extends PackagerSettings {
     @SuppressWarnings("unchecked")
     public Packager createPackager() throws Exception {
 
-        PackagePluginExtension extension = packageTask.getProject().getExtensions().findByType(PackagePluginExtension.class);
+        GradlePackagerSettings extension = packageTask.getProject().getExtensions().findByType(GradlePackagerSettings.class);
 
         return
                 (Packager) PackagerFactory
