+++ src/main/java/io/github/fvarrui/javapackager/utils/NativeUtils.java
@@ -1,0 +1,9 @@
+package io.github.fvarrui.javapackager.utils;
+
+import java.io.File;
+
+public class NativeUtils {
+    public static File getUserTempFolder(){
+        return new File(System.getProperty("java.io.tmpdir"));
+    }
+}
